moved CheckboxProps to types folder
